* UPD: Refactoring CDB_Model and CDB_OptimizeTree code.
* BACK: Restored CDB-construction thread and added the CD_BUILDER_DEBUG macro for disable this thread.
