+++ code/engine.vc2008/FrayBuildConfig.hpp
@@ -3,6 +3,9 @@
 //#define SLOW_VERIFY_ENTITIES	// Very slow verifing every entity on every game update
 //#define MEM_DEBUG				// Debbuging memory allocator
 
+/* Threads */
+//#define CD_BUILDER_DEBUG		// Binding CDB-thread to the main.
+
 /* Scripts */
 //#define LUACP_API				// attaching luaicp.dll
