+++ code/engine.vc2008/xrCDB/xrCDB.vcxproj
@@ -256,7 +256,8 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="CDB_Model.cpp" />
+    <ClCompile Include="xrCDB_Model.cpp" />
+    <ClCompile Include="xrCDB_Tree.cpp" />
     <ClCompile Include="Frustum.cpp">
       <MinimalRebuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">false</MinimalRebuild>
     </ClCompile>
@@ -298,7 +299,8 @@
     <ClCompile Include="xrXRC.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="CDB_Model.h" />
+    <ClInclude Include="xrCDB_Model.h" />
+    <ClInclude Include="xrCDB_Tree.h" />
     <ClInclude Include="cl_intersect.h" />
     <ClInclude Include="Frustum.h" />
     <ClInclude Include="StdAfx.h" />
