+++ code/engine.vc2008/xrCDB/xrCDB_Model.h
@@ -1,0 +1,17 @@
+#pragma once
+#include "xrCDB_Tree.h"
+
+class CDB_Model: public Opcode::Model
+{
+public:
+	void Store	(IWriter* pWriter);
+	void Restore(IReader* pReader);
+
+public:
+	CDB_Model	();
+	bool Build	(const Opcode::OPCODECREATE& create);
+IC  CDB_OptimizeTree*	GetTree() { return pTree; }
+
+protected:
+    CDB_OptimizeTree * pTree;
+};
