+++ code/engine.vc2008/xrCDB/xrCDB_box.cpp
@@ -2,7 +2,7 @@
 #pragma hdrstop
 
 #include "xrCDB.h"
-#include "CDB_Model.h"
+#include "xrCDB_Model.h"
 
 using namespace CDB;
 using namespace Opcode;
