+++ code/engine.vc2008/xrCDB/xrCDB_frustum.cpp
@@ -3,7 +3,7 @@
 
 #include "xrCDB.h"
 #include "frustum.h"
-#include "CDB_Model.h"
+#include "xrCDB_Model.h"
 
 using namespace CDB;
 using namespace Opcode;
